Reorganize class namespaces.
